requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy[asyncio]
asyncpg
minio
aiohttp # Для асинхронных HTTP-запросов к LLM
# Основные зависимости парсеров
marker-pdf[full] # Включает torch, surya-ocr и т.д.
python-docx
openpyxl
pygments
sensory-data-client
redis